requirements.txt
Flask
flask-cors
tensorflow
numpy
pillow
gunicorn
